org-babel-hy-evaluate-external-process fix command quoting
Replace single quotes with combine-and-quote-strings when passing
source code to org-babel-eval.

hy source blocks were passed to orb-babel-eval using a pair of single
quotes, which meant that if a quote was used in a source block it
would end the string and the rest of the source block would be fed
directly to the shell used by org-babel-eval.